commit: handler error for supplier done. next: into products.

+++ src/controllers/supplierController.js
@@ -31,6 +31,51 @@ const getSupplier = async (req, res) => {
 // Create a new supplier
 const createSupplier = async (req, res) => {
   try {
+    const { name, address, company, email, phone } = req.body
+
+    // CREATE HANDLER TO SIMPLIFY CODE
+    TODO: if (!name) {
+      return res.status(400).json({ error: 'Name is required' })
+    }
+
+    if (!address) {
+      return res.status(400).json({ error: 'Nickname is required' })
+    }
+
+    if (!company) {
+      return res.status(400).json({ error: 'Name is required' })
+    }
+
+    if (!phone) {
+      return res.status(400).json({ error: 'Phone is required' })
+    }
+
+    // CHECK IF THE PHONE EXISTES ON DB
+    const existingSupplierByPhone = await prisma.supplier.findUnique({
+      where: { phone: phone }
+    })
+
+    if (existingSupplierByPhone) {
+      return res.status(400).json({ error: 'Phone number already exists.' })
+    }
+
+    const emailRegex = /\S+@\S+\.\S+/
+    if (!email) {
+      return res.status(400).json({ error: 'Email is required' })
+    } else if (!emailRegex.test(email)) {
+      return res.status(400).json({ error: 'Email is not valid' })
+    }
+
+    // CHECK IF THE EMAIL EXISTES ON DB
+    const existingSupplierByEmail = await prisma.supplier.findUnique({
+      where: { email: email }
+    })
+
+    if (existingSupplierByEmail) {
+      return res.status(400).json({ error: 'Email already exists.' })
+    }
+
+    // IF PASSES THROUGH VALIDATION, CREATE SUPPLIER
     const newSupplier = await prisma.supplier.create({
       data: req.body
     })
@@ -45,6 +90,12 @@ const createSupplier = async (req, res) => {
 const updateSupplier = async (req, res) => {
   try {
     const { id } = req.params
+    const supplier = await prisma.supplier.findUnique({
+      where: { id: parseInt(id) }
+    })
+
+    if (!supplier) return res.status(400).send('Supplier not found')
+
     const updatedSupplier = await prisma.supplier.update({
       where: { id: Number(id) },
       data: req.body
@@ -56,10 +107,18 @@ const updateSupplier = async (req, res) => {
   }
 }
 
+
 // Delete a supplier
 const deleteSupplier = async (req, res) => {
   try {
     const { id } = req.params
+
+    const supplier = await prisma.supplier.findUnique({
+      where: { id: parseInt(id) }
+    })
+
+    if (!supplier) return res.status(400).send('Supplier not found')
+
     await prisma.supplier.delete({
       where: { id: parseInt(id) }
     })
@@ -68,7 +127,6 @@ const deleteSupplier = async (req, res) => {
     console.error(error)
     res.status(500).send('An error occurred while deleting the supplier')
   }
-  // add validator to see if user exists or not and send error
 }
 
 module.exports = {
